fix time display for time entries in report email

+++ src/main/java/org/manage/service/ReportService.java
@@ -31,7 +31,7 @@ public class ReportService {
         dayReportDTO.date = reportDate;
         dayReportDTO.membersReports = membersReports;
         dayReportDTO.project = projectDto;
-        dayReportDTO.totalHours = dayReportDTO.membersReports.stream().collect(Collectors.summingLong(r -> r.totalHours));
+        dayReportDTO.totalHours = dayReportDTO.membersReports.stream().collect(Collectors.summingDouble(r -> r.totalHours));
         dayReportDTO.subject = "Отчёт по работе над проектом " + dayReportDTO.project.name + " за " + LocalDateProvider.dateFormatter.format(reportDate);
         return dayReportDTO;
     }
@@ -40,8 +40,8 @@ public class ReportService {
         MemberReportInfoDTO dto = new MemberReportInfoDTO();
         dto.member = m;
         dto.entries = timeEntryService.findByMemberAndDateAndProject(m.id, reportDate, projectId);
-        Long totalSeconds = dto.entries.stream().collect(Collectors.summingLong(e -> e.duration.toSeconds()));
-        dto.totalHours = totalSeconds / 60 / 60;
+        Double totalSeconds = dto.entries.stream().collect(Collectors.summingLong(e -> e.duration.toSeconds())).doubleValue();
+        dto.totalHours = totalSeconds / 60.0 / 60.0;
         return dto;
     }
